backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
apscheduler
openai
python-dotenv
aiofiles
google-generativeai
